@@ -1,15 +1,16 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
+import ReactDOM from 'react-dom/client';
 import App from './App';
-import { BrowserRouter } from "react-router-dom"
-ReactDOM.render(
-  // <React.StrictMode>
-  <BrowserRouter>
-    <App />
-  </BrowserRouter>,
+import 'antd/dist/antd.min.css'
+import { Provider } from "react-redux"
+import store from './store';
+const root = ReactDOM.createRoot(document.getElementById('root'));
 
-  //</React.StrictMode>,
-  document.getElementById('root')
+root.render(
+  // <React.StrictMode>
+    <Provider store={store}>
+      <App />
+    </Provider>
+  // </React.StrictMode>,
 );
 
