@@ -1,9 +1,13 @@
+import React from 'react';
 import './App.css';
-
+import { BrowserRouter, Routes, Route } from "react-router-dom"
+import IndexRouter from './router/IndexRouter';
 function App() {
   return (
     <div className="App">
-      123
+      <BrowserRouter>
+        <IndexRouter />
+      </BrowserRouter>
     </div>
   );
 }
